// backend/src/modules/production-orders/dto/create-production-order-semi-finished-net-weight-check.dto.ts
export class CreateProductionOrderSemiFinishedNetWeightCheckDto {
  requirement?: string | null;
  unit_1_net_weight?: number | string | null;
  unit_2_net_weight?: number | string | null;
  unit_3_net_weight?: number | string | null;
  unit_4_net_weight?: number | string | null;
  unit_5_net_weight?: number | string | null;
  unit_6_net_weight?: number | string | null;
  unit_7_net_weight?: number | string | null;
  unit_8_net_weight?: number | string | null;
  unit_9_net_weight?: number | string | null;
  unit_10_net_weight?: number | string | null;
  unit?: string | null;
}